Separate django files
