--- Dockerfile
@@ -1,26 +1,0 @@
-FROM alpine
-
-# init
-RUN mkdir /www
-WORKDIR /www
-COPY requirements/ /www/requirements
-
-# setup
-RUN apk update
-RUN apk upgrade
-RUN apk --no-cache add \
-    python3 \
-    python3-dev \
-    postgresql-client \
-    postgresql-dev \
-    build-base \
-    gettext
-RUN pip3 install --upgrade pip
-RUN pip3 install -r requirements/production.txt
-
-# clean
-RUN apk del -r python3-dev postgresql
-
-# prep
-ENV PYTHONUNBUFFERED 1
-COPY . /www/
