+++ django/config/settings/development.py
@@ -1,0 +1,12 @@
+import os
+from .base import *
+
+DEBUG = True
+
+# Database
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.sqlite3',
+        'NAME': os.path.join(BASE_DIR, 'db.sqlite3'),
+    }
+}
