+++ django/config/settings/production.py
@@ -1,0 +1,20 @@
+import os
+from .base import *
+
+DEBUG = bool(int(os.getenv('DEBUG', False)))
+
+ALLOWED_HOSTS = os.environ['ALLOWED_HOSTS'].split(',')
+
+WSGI_APPLICATION = 'config.wsgi.application'
+
+# Database
+DATABASES = {
+    'default': {
+        'ENGINE': 'django.db.backends.postgresql_psycopg2',
+        'HOST': os.getenv('POSTGRES_HOST'),
+        'PORT': os.getenv('POSTGRES_PORT'),
+        'USER': os.getenv('POSTGRES_USER'),
+        'PASSWORD': os.getenv('POSTGRES_PASSWORD'),
+        'NAME': os.getenv('POSTGRES_DB'),
+    }
+}
