+++ django/recite/apps.py
@@ -1,0 +1,5 @@
+from django.apps import AppConfig
+
+
+class ReciteConfig(AppConfig):
+    name = 'recite'
