+++ django/requirements/development.txt
@@ -1,0 +1,1 @@
+-r base.txt
