+++ docker-compose.yml
@@ -1,14 +1,14 @@
 version: "3"
 services:
   web:
-    build: .
+    build: ./django/
     container_name: web
     env_file:
       - ./.env
     entrypoint: "/www/entrypoint.sh"
     command: "gunicorn config.wsgi -b 0.0.0.0:8000 --workers 3 --timeout 600"
     volumes:
-      - .:/www
+      - ./django/:/www
     ports:
       - "8000:8000"
     depends_on:
@@ -39,8 +39,8 @@ services:
     image: nginx:latest
     volumes:
       - ./nginx:/etc/nginx/conf.d
-      - ./media:/var/www/quran-words/media
-      - ./collected_static:/var/www/quran-words/collected_static
+      - ./django/media:/var/www/quran-words/media
+      - ./django/collected_static:/var/www/quran-words/collected_static
     ports:
       - "80:80"
     depends_on:
