--- requirements/development.txt
@@ -1,1 +1,0 @@
--r base.txt
